requirements.txt
python-dateutil~=2.9.0.post0
pandas~=2.2.3
pyperclip~=1.9.0
matplotlib~=3.9.4